services/auth-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pyjwt>=2.8.0
pydantic>=2.0.0
